new/requirements.txt
requests
termcolor
sockets
python-whois
beautifulsoup4
pyautogui